+++ v4/export/black_white_list_test.go
@@ -1,8 +1,11 @@
 package export
 
 import (
+	"strings"
+
 	. "github.com/pingcap/check"
 	"github.com/pingcap/tidb-tools/pkg/filter"
+	"github.com/pingcap/tidb/util"
 )
 
 var _ = Suite(&testBWListSuite{})
@@ -37,4 +40,43 @@ func (s *testBWListSuite) TestBWList(c *C) {
 		},
 	})
 	c.Assert(err, NotNil)
+}
+
+func (s *testBWListSuite) TestFilterTables(c *C) {
+	dbTables := DatabaseTables{}
+	expectedDBTables := DatabaseTables{}
+
+	dbTables.AppendTables(util.InformationSchemaName.L, []string{"xxx"}...)
+	dbTables.AppendTables(strings.ToUpper(util.PerformanceSchemaName.L), []string{"xxx"}...)
+	dbTables.AppendTables("xxx", []string{"yyy"}...)
+	expectedDBTables.AppendTables("xxx", []string{"yyy"}...)
+	dbTables.AppendTables("yyy", []string{"xxx"}...)
+
+	conf := &Config{
+		ServerInfo: ServerInfo{
+			ServerType: ServerTypeTiDB,
+		},
+		Tables: dbTables,
+		BlackWhiteList: BWListConf{
+			Mode: OldToolsMode,
+			OldTools: &OldToolsConf{
+				Rules: &filter.Rules{
+					DoDBs: []string{""},
+				},
+			},
+		},
+	}
+
+	c.Assert(filterTables(conf), NotNil)
+	conf.BlackWhiteList = BWListConf{
+		Mode: OldToolsMode,
+		OldTools: &OldToolsConf{
+			Rules: &filter.Rules{
+				DoDBs: []string{"xxx"},
+			},
+		},
+	}
+	c.Assert(filterTables(conf), IsNil)
+	c.Assert(conf.Tables, HasLen, 1)
+	c.Assert(conf.Tables, DeepEquals, expectedDBTables)
 }
