+++ v4/export/prepare.go
@@ -5,8 +5,6 @@ import (
 	"strings"
 
 	"github.com/pingcap/dumpling/v4/log"
-	"go.uber.org/zap"
-	"github.com/pingcap/tidb/util"
 )
 
 func detectServerInfo(db *sql.DB) (ServerInfo, error) {
@@ -93,42 +91,4 @@ func (d DatabaseTables) Merge(other DatabaseTables) {
 	for name, infos := range other {
 		d[name] = append(d[name], infos...)
 	}
-}
-
-func filterDirtySchemaTables(conf *Config) {
-	switch conf.ServerInfo.ServerType {
-	case ServerTypeTiDB:
-		for dbName := range conf.Tables {
-			if filter.IsSystemSchema(strings.ToLower(dbName)) {
-				log.Zap().Warn("unsupported dump schema in TiDB now", zap.String("schema", dbName))
-				delete(conf.Tables, dbName)
-			}
-		}
-	}
-}
-
-func filterTables(conf *Config) error {
-	log.Zap().Debug("filter tables")
-	// filter dirty schema tables because of non-impedance implementation reasons
-	filterDirtySchemaTables(conf)
-	dbTables := DatabaseTables{}
-	bwList, err := NewBWList(conf.BlackWhiteList)
-	if err != nil {
-		return withStack(err)
-	}
-
-	for dbName, tables := range conf.Tables {
-		doTables := make([]string, 0, len(tables))
-		for _, table := range tables {
-			if bwList.Apply(dbName, table.Name) {
-				doTables = append(doTables, table.Name)
-			}
-		}
-		if len(doTables) > 0 {
-			dbTables = dbTables.AppendTables(dbName, doTables...)
-		}
-	}
-
-	conf.Tables = dbTables
-	return nil
 }
